requirements.txt
absl-py==2.3.1
anyio==4.12.0
astunparse==1.6.3
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
dlib==20.0.0
exceptiongroup==1.3.1
face-recognition==1.3.0
face_recognition_models==0.3.0
flatbuffers==25.9.23
gast==0.7.0
google-pasta==0.2.0
googletrans==4.0.2
grpcio==1.76.0
gTTS==2.5.4
h11==0.16.0
h2==4.3.0
h5py==3.14.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
joblib==1.5.2
keras==3.10.0
libclang==18.1.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
nltk==3.9.2
numpy==2.0.2
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pillow==11.3.0
protobuf==6.33.2
PyAudio==0.2.14
Pygments==2.19.2
regex==2025.11.3
requests==2.32.5
rich==14.2.0
six==1.17.0
soupsieve==2.8
SpeechRecognition==3.14.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.4
wikipedia==1.4.0
wrapt==2.0.1
zipp==3.23.0
